requirements.txt
deep_translator==1.11.4
Flask==3.0.1
SocksiPy_branch==1.01
